// frontend/src/api/models/UserDeliveryDataRequestRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserDeliveryDataRequestRequest = {
	full_name?: string;
	email?: string;
	phone?: string;
	zip_code?: string;
	address_line?: string;
	destination_code?: string;
	region_code?: string;
};