// frontend/src/api/models/UserDeliveryDataResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserDeliveryDataResponse = {
	full_name?: string;
	email?: string;
	phone?: string;
	zip_code?: string;
	address_line?: string;
	destination_code?: string;
	region_code?: string;
	id: number;
	is_current?: boolean;
};